Add .gitignore file and improve image resizing script
.gitignore file is added to ignore non-essential generated files which should not be tracked by GIT.

In 'main.py', made amendments to homogenize filenames and handle file existence errors for thumbnail directory function. Also improved readability of the code and reduced the redundancy in the code.

In '.idea/misc.xml', updated the 'component' parameters to better suit the project requirements.

In '.idea/vcs.xml', added a mapping directory which now uses GIT as the version control system.

These changes are made to streamline the code and versioning process.